New Gravity Wave python notebook, Occultation Package notebook and .py file and a modified version of the Stellar Occultation python notebook.
The gravity wave notebook holds all of my (Pepper) code for this project and is focused on using Amelias code (Stellar Occultation) to find gravity waves. The Stellar Occultation notebook was updated to use relative paths and converted over into the Occultation Package notebook and .py file which contains the functions she built last year, the functions are a regular and smoothed column density function, an irradiance function and number density function. Everything is now path relative as opposed to using hard-coded paths.