Added Temperature&Column-Density notebook, and changed the main Gravity Wave Notebook
The Gravity Wave main code has been updated to smooth the column density function at the extinction ratio level and it now used the extinction ratio rather than a hard-coded height range, to determine what part of the data should be looked at and have the line fitted over top it. The Temperature&Column-Density notebook was added to test the amplitude of column density perturbations relative to some temperature perturbations.